Use MAP_ANON when MAP_ANONYMOUS not available
This commit is a temporary requirement for JitBuilder and other
OMR compiler based projects to be able to build on older versions
of OS X where MAP_ANONYMOUS has not been defined. This problem
came up in Issue #902 while building JitBuilder and the OMR compiler
component on Mac OS X 10.9 . In these kinds of environments, MAP_ANON
is usually defined and so we can simply define MAP_ANONYMOUS to
be MAP_ANON where we need it. The port library already does this
swizzling but we have not yet hooked up a depenedency for the
compiler or JitBuilder on the port library to benefit from that
code.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>