Add runtime flag for concurrent scavenge support
Concurrent scavenger related code enabling the code generator to emit
read barriers for object reference loads is currently guarded by a
compile time OMR_GC_CONCURRENT_SCAVENGER define.

This change effectively deprecates the OMR_GC_CONCURRENT_SCAVENGER flag
in the JIT in favor of a runtime check. This is a pre-emptive change in
preparation of full deprecation of the OMR_GC_CONCURRENT_SCAVENGER flag.
However this removal has to come from the OMR GC front and a
corresponding runtime flag for concurrent scavenger needs to be
introduced in the GC. Only at this point may we remove the final
reference to OMR_GC_CONCURRENT_SCAVENGER in the JIT.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>