Fix illegal cp option in JitBuilder makefiles
On Ubuntu, you can use cp -u to only copy files that have changed,
which is handy when you're copying header files that may not change
from build to build and you don't want to trigger things to be
rebuilt if they don't need to be rebuilt.

Alas, the cp command on OS X does not believe in this option and
prints a message that makes it look a lot like an error has occurred.
The problem was mentioned in Issue #902 as part of that experience
report building the JitBuilder library.

For simplicity, I'm just going to remove it to eliminate that message.

Only people who regularly build and rebuild the JitBuilder library
will be affected though probably not terribly severely.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>