Use a counter for isEvaluatingCompressionSequence
Use a counter to keep track of whether we are currently evaluating a
compression sequence or not. We use a counter to avoid a scenario under
which multiple compression sequences could be nested underneath each
other in which case we may skip to set the appropriate codegen flag.

Not marking compression sequences during codegen could result in missed
guarded loads for concurrent scavenge.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>