Extend remat API to analyse beyond one block
The remat API has several functions that require a start and end
node to be specified. In this form, they are limited to a block
or a continuous set of blocks, as it is not possible to specify
which edges should be taken. This change extends the API to include
a BitVector of allowable blocks. When these walks reach the
end of their current block, the first allowable block in the
set of successors will be taken, allowing for more aggressive
remat. To ensure the path is deterministic, none of the blocks
along the path can be a merge point except for the start.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>