Remove vestigial RTSJ comments from compiler
A long time ago, the compiler component included a whole slough
of adaptations so that it could operate in real-time environments
and support the Real Time Specification for Java (RTSJ). When
that code was initially introduced, it was bracketed by comments
indicating which parts were specifically for RTSJ or real-time.
Support for the RTSJ has been removed at this point (though much
of the code lives on and operates for all languages as part of
the OMR compiler), but there were some vestigial comments still
lying around identifying code as related to the RTSJ. This commit
removes these comments as they are no longer appropriate or
useful.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>