Enable on-demand method lookup for JitBuilder
Currently if JitBuilder clients want to Call a function using either
Call or ComputedCall, that function must have been described to
JitBuilder via the DefineFunction call on the MethodBuilder object.
All Calls lookup the function name provided and will fail to call it
if DefineFunction has not yet been called. The only way to make sure
the function has been defined is to make sure you call it yourself
before that function name is called.

This commit adds one check inside the Call services so that, if the
first lookup fails, it will call a virtual instance on MethodBuilder
called RequestFunction. Clients can implement this instance function
in their subclass of MethodBuilder to locate and call DefineFunction
for the requested function name. When RequestFunction() comes back,
the Call service will try to look the function up again and will
proceed if it has now been defined. This approach frees JitBuilder
clients from having to define all possible functions up front, which
may not be practical. Instead, required functions can be defined
on demand.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>