Fix JitBuilder broken recursive method name test
Hugely silly test to check for a recursive call to the function
being compiled only used strncmp with the length of the current
function without also checking that the names have the same
length. So looking up "mandelhelp" would actually short circuit
the lookup if you were compiling "mandel" because, well, the
substring matched :( . Unanticipated recursions ensue only to be
discovered as the 75,000th deep call blows the stack...

This commit adds a first check that the names have the same
length before using strncmp with one of the string's lengths.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>