Remove the initialized check on a class
Loosen the condition on constraining indirect load from a J9Class by
removing the initialized check. This is because java/lang/Class
reference and the component type if the class is an array, are
properties of the class file, and will be initialized before the class
is sufficiently initialized, so it's safe to dereference them at
compile time without the initialized check.

Signed-off-by: liqunl <liqunl@ca.ibm.com>