Add common functions to OSR API concerning PPS
OSR points are surrounded by pending push stores, to ensure
the stack is recreated after the transition, and treetop of
pending push stores, to avoid the side effects of the prior mentioned
stores. This adds functionality to determine if a node is one of
these and to extract the transition point for an OSR point based
on the current mode.

Moreover, it corrects an issue with OSR analysis as liveness and
use/def information should be calculated at the transition or
analysis point, not the OSR point.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>