Extend OSR API to represent its different modes
This change adds checks that abstract over the different OSR
modes, specifically whether it is operating in voluntary
or involuntary OSR mode and whether it will insert the OSR
transition before or after the OSR point.

Moreover, it adds the concept of OSR analysis points, which
exist only to hold OSR information at certain bytecode indices and will
not become a transition point.

A further change is partial de-aliasing of the osrInfrastructureRemoved
and the canAffordOSRControlFlow flags. canAffordOSRControlFlow is used
to indicate whether the additional compile time of OSR is afforadable,
whilst osrInfrastructureRemoved indicates that OSR exception edges
have been removed. However, canAffordOSRControlFlow was being used
to achieve the latter in some cases.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>