Add support for VirtualMachineOperandArray
Not all Virtual Machines have an Operand Stack. Another common construct
is to have a fixed sized array of "registers". The new
VirtualMachineOperandArray is very similar to the VMOS but it does not
allow for growth and it has to deal with NULL entries as not all entries
may have a value.

The VirtualMachineOperandArray::Move API requires that a copy of the
IlValue stored at srcIndex be stored at dstIndex. To accomplish this I
added a new API to IlBuilder call Copy.

I have added tests to verify the all of the functionality.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>