Support MethodBuilders without line or file info
Before this patch, if you forgot to add DefineLine() or DefineFile() to
your MethodBuilder (perhaps because you don't have that information
available), JitBuilder would crash during compilation as it tried to
compute the signature of the method.

This tweaks MethodBuilder to allow MethodBuilders without line or files
by just using an empty string if they're not available. A test is also
added to ensure this functionality continues to work. The test creates
an anonymous method builder, and simply returns a value.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>