Notify CM when old-to-old ref is created by CS
When Concurrent Scavenger (CS) creates a new old-to-old reference it has
to notify Concurrent Marker (CM). Otherwise if parent object is already
scanned, the child object will be collected.

CM will dirty the card of the parent object (if marked, since we have no
information if it's scanned), what will eventually through card cleaning
find the child object.

This is generalization of the functionality where regular Scavenger
notifies CM about removed RS slot, which is the case where old-to-old
reference was created from old-to-new reference, when child object is
tenured (thus parent removed from RS). 

In CS, even new-to-new reference is interesting, when both parent and
child object are tenured and old-to-old reference is created. While CS
is in progress, it may overlap with CM in progress, so CM can mark and
scan the parent object BEFORE the child object is tenured.

This last scenario is possible in regular Scavenger, except, since the
world is stopped during Scavenge, the parent object is guaranteed to be
NOT marked/scanned (CM concurrently traces only old objects, and is
stopped). 

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>