Expand AutoSIMD support on z to floats
Updates the implementation of the following evaluators on z to support the
float data type:
vadd, vsub, vmul, vdiv, vneg
vsplats, getvelem, vsetelem

Many of those evaluators made calls to inlineVectorUnaryOp and
inlineVectorBinaryOp. Those functions were also updated to handle the
float data case. They now set the proper bits in a vector instruction when
a float version is used.

In most of the modified functions, a check for zNext was added since it is
necessary for the float versions of the vector instructions to be used.

getSupportsOpCodeForAutoSIMD was updated to reflect the new support for
floats on zNext.

Added comment to getSupportsOpCodeForAutoSIMD explaining how zNext is
required for AutoSIMD float support.

Renamed variable 'v' in vsplatsEvaluator on z to "value" to be more
descriptive and to match similar code in vsetelemEvaluator.

Issue: #950
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>