Array Compare Evaluator Improvements
Two improvements were implemented:

1. Taking Advantage of REPE CMPS Instructions

The REPE CMPS Instructions can be used to keep comparing two arrays
until an inequality is found. Furthermore REPE can be paired with CMPSQ,
CMPSD, CMPSW or CMPSB to compare 8, 4, 2 or 1 byte(s) respectively. In
this implementation, REPE CMPSQ is used on 64bit machines, while REPE
CMPSD is used on 32bit machines. The residual bytes are dealt with a
sequence of CMPSD, CMPSW, CMPSB instructions. If it is neccessary to
determine the specific byte at which the inequality occurs (arrayCmpLen)
then REPE CMPSB is used.

2. Constant Length Array Comparison Sequence

If the length of the array is known at compile time, and this length
falls within a specific threshold, then a "Constant Length Optimization"
can be performed. This will generate a sequence of CMPSQ (64bit machine)
or CMPSD (32bit machine) that will compare 8/4 bytes at a time, taking
care of residuals with CMPSD, CMPSW and CMPSB or CMPSB depending on
arrayCmp or arrayCmpLen.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>