Templatize compileOpCodeMethod
By doing this we can remove the casts from callers of this method, and
dramatically improve readability. compileOpCodeMethod is modified to
take a reference to the target function pointer, rather than return it.
This allows the template types to be inferred.

To allow implicit expansion of the template function, the implementation
of compileOpCodeMethod moves into the OpCodeTest header file.

Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>