Teach the linter about most derived type calls
The linter needs to catch calls that are missing downcasts. What happens
however when the 'downcast' would really resolve to the current type?
This is a case we have encountered in the codebase where a most-derived
concrete TR:: type makes a member function call (see PR#913).

This commit resolves this by teaching the linter that calls to the most
derived type, where the caller is already the most derived type, are OK.

Resolves #915
Signed-off-by: Matthew Gaudet <magaudet@ca.ibm.com>