Fix block split assert
There is an assert in the edge splitting code which asserts a block
connected to the exit must end in a throw or a return. It checks this
by looking at the treetop. These nodes are allowed to appear under check
treetops and this change fixes the assert to correctly handle these
cases.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>