Add infra to save basic block offset and frequency
Reassociating profiling information can be very tricky. In general the
OMR compiler will use bytecode metadata stored alongside profiling data
to faciliate re-associating the profiling data with basic blocks. It is
general practice in the compiler to reset all block frequencies post
inlining to regenerate a re-nomalized set of frequency data for the
composed method as a whole. This means we can see blocks which have been
split by inlining and we need to find the correct bytecode index at
which to lookup their frequency data. It is also possible that we are
not able to recompute a frequency for a block even though a frequency
was computed for it when it was IL generated in isolation. For this
reason it would be desirable to save frequency information from il
generation as a fallback. This change adds the infrastructure to save
bytecode ranges for basic blocks as seen by IL generation and their
associated frequencies for later use during profile re-association.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>