Add initial infrastructure for new JProfiling
Now that the old JProfiling implementation based on hardware sampling
has been renamed to SamplingJProfiling, I would like to add initial
placeholder infrastructure to support the new deterministic JProfiling
that is in development in a downstream project. The plan is to
contribute this new implementation, but I need to do additional
performance verification and tuning first as well as some refactoring.
This change adds minimal placeholders to make it easier to contribute
the new implementation in the (hopefully near) future.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>