Add check for OSR induce block
When OSR transitions are implemented using OSR guards, the taken
side of the guard manages several book keeping aspects of the
transition. For example, it will contain stores to dead symbol
references and will remat symbol references required for the
transition. This block can be ignored by several optimizations.
To allow for this, this change adds a check for such a block.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>