Ensure numa.h is available if required
Modify the OMR_PORT_NUMA_SUPPORT configure define to check for numa.h.
If it is --enabled and not found print an error message.

Only set --enable-OMR_PORT_NUMA_SUPPORT for one of the Linux x86
platforms. We do not need to test this in all specs.

Update INSTALL to remove the requirement for numa-dev

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>