Inhibit array splitting in Scavenger RS overflow
Inhibit array splitting in MM_Scavenger remembered set processing when
remembered set is in overflow state.

Also removed check for backout condition in scavenger getObjectScanner()
inline method as the flag to inhibit array splitting is applied in any
case in shouldRememberObject(), so this check is now redundant. The 
only calls to shouldRemeberObject() are from contexts where backout 
flag is set or rememebered set is in overflow state.

Signed-off-by: Kim Briggs <briggs@ca.ibm.com>