Re-implement ARM binary encoding
The old implementation of `ARM::CodeGenerator::doBinaryEncoding` is
very Java specific. This implementation is therefore moved into J9
and an alternate, generic implementation is provided as a substitute.

Also, part of the binary encoding mechanism for branch instructions is
guarded as it too is very Java specific. An alternate, generic
implementation is provided as well.

[ci skip] because no arm testing is currently available

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>