Specialized 'fe' option processing routines
Provide specialized functions for option processing functions that need
to operate on JitConfig. The JitConfig is actually cached in a static
field in OMR::Options::_feBase, so it is readily available.
This change paves the way to splitting the option flags into OMR generic
ones and option flags that are only used by subprojects.

Issue: #1055
Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>