Fix fetching the compilationDirectory attribute
We currently are missing a call to dwarf_attr() to fetch the
DW_AT_comp_dir attribute.  This call was there previously, but lost in
change 055c3bfe8c2356c6bfb9074c896a219264947c51.  This means that attr
will always be NULL.

On my Linux machine, dwarf_formatstring() will segfault when passed attr
with a value of NULL.  The manual states that dwarf_formatstring() will
return DW_DLE_ARGUMENT if "Either of arguments or ret was NULL", but
also states "The form of argument attr must be one of DW_FORM_string or
DW_FORM_strp".

Signed-off-by: Andrew Young <youngar17@gmail.com>