Fix live registers issue for concurrent scavenge
For Guarded Storage running with compressed refs shift > 0, we are not
evaluating shift node since the result of guarded load instruction is
already shifted. We need to decrement ref count on a child of shift
node. This code will get removed once we deliver a proper fix to not
expand compressed refs in lower trees when concurrent scavenge is
enabled.

Signed-off-by: evgeniab@ca.ibm.com <evgeniab@ca.ibm.com>