Add MVC support for reference array copy
Currently the z Systems codegen is generating a load-store loop for
forward reference arraycopy because of a historic bug that is now moot.
This commit enables generation of the same MVC sequence for reference
arraycopy that we use for primitive arraycopy.

We take this time to reorganize, simplify, and document all arraycopy
evaluators on z Systems. We now have a consolidated path for primitive
and reference arraycopy with maximal code reusage.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>