Modify JitBuilder Switch sample to use RequestFunction
Arbitrarily chose the Switch code sample from the jitbuilder/release
directory to show how MethodBuilder subclasses can use RequestFunction
to lazily call DefineMethod only for methods that are actually
referenced (Called) via ::buildIL(). While this facility is more
likely to be used in a JIT compiler scenario where the MethodBuilder
constructor may not know ahead of time which methods will actually be
called while processing the method's operations, there are not yet
any JitBuilder code samples like that. Switch is a simple code
sample that only calls one external function, so it can serve
as a simple demonstration of how to use RequestFunction. There
are many other samples that demonstrate how to call DefineFunction
inside the constructor.

The actual change calls DefineFunction inside the virtual
RequestFunction() implementation rather than as part of the
constructor. When the buildIL() method issues the first Call to
"printString", because that function has not yet been defined,
RequestFunction will be called and that's how "printString" will
get defined.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>