Clean up third-party usage of TR::CFGEdgeList
There are quite a few locations in the codebase where various parts of
the JIT [rightly] interact with TR::CFGEdgeLists. A significant number
of those locations, however, make assumptions about the implementation
details of TR::CFGEdgeList. This change fixes the locations found to be
problematic.

One of the primary motivating factors was to avoid unnecessary copying
of TR::CFGEdgeList. For most locations, this could be improved by using
a TR_CFGIterator rather than doing direct copies / manipulations. How-
ever, copying was more systemic in a couple of locations: In induction
variable analysis, we were using a private instance of TR:CFGEdgeList as
a work. In loop reducer, we were passing TR::CFGEdgeList objects by
value.

For TR_InductionVariableAnalysis, we redefined the work queue object
type to use std::deque rather than TR::CFGEdgeList.  This had the fur-
ther benefit of allowing this work queue to use an independent alloca-
tion strategy.

For TR_LoopReducer, we eliminate redundancy with OMR::CFG. We localize
removeSelfEdge to TR_LoopReducer, and shed the local implementation of
removeEdge. Secondly, we modify the function signature such that each
TR::CFGEdgeList is passed by reference, rather than by value. Finally,
we turn instances of automatic scope into references, having reviewed
each case and determined that [as far as we can tell] there are no
iterators outstanding that risk invalidation.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>