Allow late decision whether to obey CS request
Concurrent Scavenger: Introduce an additional 'requested' flag that will
be set during parsing of the CS command line option. There might be
platform/language specific requirements after we parsing is done that
will dictate whether the request is obeyed or not.

Add a line in 'initialized' verbose GC stanza to report if we actually
run CS or not.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>