Add flag to skip setting frequency
Currently, frequencies will be calculated and set before an
optimization when all necessary information is present. It is then
maintained across modifications to the CFG. This necessary information
includes structure analysis.

This change allows an optimization to explicitly prevent frequencies
from being set before itself. This is useful for early opts that require
structure, but not frequency, and may significatly alter the structure
to the point where the frequency data would not have been adequately
maintained.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>