Fix JitBuilder StructFieldInstanceAddress service
On 32-bit platforms, the offset of a field must be specified as a 32-bit
constant so that adding the value to a struct base address is
type-consistent. This commit re-uses the same approach used to
initialize the `Word` field in OMR::TypeDictionary to ensure the offset
is of the correct size on a given platform.

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>