Use consistent exclusive access backoff
The tryAcquireExclusiveVMAccess path in the GC will wait for another
thread if it is already acquiring exclusive VM access.  It back off
using releaseCriticalHeapAccess and reacquireCriticalHeapAccess.  The
acquireExclusiveVMAccess uses releaseVMAccess(bool) and acquireVMAccess.
Change acquireExclusiveVMAccess to be consistent with
tryAcquireExclusiveVMAccess so the delegate may more easily detect the
exclusive backoff case.

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>