Fix a race in ThreadAbortTest.Sleeping
In the original test, the sleeper could enter and notify on enterSync
before main hits wait.  This could result in the main thread waiting
for a notify that never comes.

The new test also replaces a hardcoded sleep in main with a notify. In
an ideal world, it would be safe to call abort as soon as
omrthread_create returns.

This probably fixes issue #1050, where this test was hanging. It now
takes ~1second on my laptop.

Signed-off-by: Robert Young <rwy0717@gmail.com>