Enforce reference arraycopy has type TR::Address
Reference arraycopies are important when concurrent scavenge is enabled.
We need to determine at code generation time that a particular arraycopy
is indeed a reference arraycopy such that we can generate read barriers
for every array element access. As such it is important that all
arraycopy nodes that are in fact reference array copies specify the data
type of the array elements is a TR::Address.

We will use this fact in the code generator to determine whether we need
to generate read barriers for array reference loads.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>