Enable Array Copy under Concurrent Scavenge
This commit enables the previously disabled (under concurrent scavenge)
transformation of the array copy call into the arraycopy node.

In order to handle a current scavenge cycle, the compiler generates code
to check a runtime flag. Execution conditionally branches to an Out Of Line
sequence which contains the necessary guarded load, shift, and store
instructions.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>