Disabled Inlined Array Copy on Power
When the array length is not known at compile time, a sequence of close
to 300 instructions is needed to cover all cases of the array copy.
Inlining this is very costly in terms of performance for small arrays.
Given that most of the arrays seen are a few dozen bytes, it is not
worth inlining. When the length of the array is known at compile time,
inlining can still occur.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>