Enable AutoSIMD support on x86
Updated getSupportsOpCodeForAutoSIMD in x/codegen/OMRCodeGenerator.cpp to
reflect the recently implemented vector evaluators. Instead of returning
false in all cases, getSupportsOpCodeForAutoSIMD will now return true for
loads, stores and vector splats for int32, int64, float and double data
and also double type add and multiply. All of these operations are
supported in the new vector evaluators.

A call to setSupportsAutoSIMD is now made inside initialize in
OMRCodeGenerator.cpp to indicate that autoSIMD is now supported. This call
is behind a check for SSE2 support.

The function isFPRUsedAsVRF and any references to it were removed. It was
added to prepareRegistersForAssignment in error. It is not needed for
AutoSIMD and sometimes caused registers to not be assigned properly. It
was only used in one place, so the entire fuction is just being removed.

Issue: #1066
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>