Assign the GPR when reverse spilling in OOL path
When reverse spilling in an out of line (OOL) path we must actually
assign the GPR to the spilled virtual register, and not only the HPR.
The reason we can do this is because if a register does not have a
backing storage in an OOL path then it must not have been ever spilled
(because it doesn't have a backing storage) which must mean the register
was first seen in the fast path of the OOL internal control flow.

As such because the register was not spilled all the way up until the
branch to the OOL path it means it must still be in a register at the
entry of the OOL path, thus it is always safe to assign the register to
the requested target real register and let the dependency on the OOL
entry label generate the proper shuffle from the main line.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>