Set arraycopy type in correct order
Compressed and decompressed array copies are also primitive array copies
and as such we must set the type and element size of these special array
copies after setting the type and element size as derived from the fact
that they are primitive array copies as well.

This is done to enforce the fact that for the specially recognized
decompressed array copies we have type set to TR::Int16 and element size
derived from the respective type.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>