Extend isPotentialOSRPoint to find OSR node
isPotentialOSRPoint will now account for monents anchored under
treetops. It will also find the OSR node within the tree. This
is used throughout the rest of the API to determine what type
of OSR point is being handled. Failure cases have also been
added to the other API calls if the OSR point is not known.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>