Cleanup and fix GC ExclusiveVMAccess APIs
There is a lot of duplicated code and makes modifications very
difficult. I am removing all of the APIs except 1 to keep things as
simple as possible.

The GC ExclusiveVMAccess APIs were not properly handling the case where
a system collect was called by a thread that already has
ExclusiveVMAccess outside of the GC. This could easily lead to a dead
lock because of the way the GC handles getting ExclusiveVMAccess. The
change also fixes the potential hang.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>