Do not anchor aloadis whos child is null checked
An aloadi cannot be anchored if there is a Null Check on
its child. There are two situations where this occurs.
The first is when doing an aloadi off some node that is
being NULLCHK'd (see Ex1). The second is when doing an
icalli in which case the aloadi loads the VFT of an
object that must be NULLCHK'd (see Ex2).

```
Ex1:
   n1n NULLCHK on n3n
   n2n    aloadi f
   n3n       aload O

Ex2:
   n1n NULLCHK on n4n
   n2n    icall foo
   n3n       aloadi <vft>
   n4n          aload O
```

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>