Fix processing of verbose options
This commit fixes a bug introduced by PR #1007.
The new implementation of setVerboseBitsInJitPrivateConfig() is able to find
the JitConfig without looking at the value of parameter 'base'. However,
when trying to set the verbose flags in PrivateJitConfig for the trj9
subproject it uses setVerboseBits() which, similarly, was changed to
ignore the parameter 'base' and retrieve the OMR JitConfig from _feBase.
This is wrong.
The solution is to factor out the commong code that sets verbose bits
and call it from two different places:
1) setVerboseBitsInJitPrivateConfig() used by trj9 subproject
2) setVerboseBits() used by OMR generic code

Issue: #1085
Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>