Split block containing the call under OSR
In any OSR mode, blocks containing OSR points should have exception
edges to the method's OSRCatchBlock. However, when inlining and
splitting a block in the caller, the callee's final block
and the second half of the split block may be merged, preventing
the addition of an edge to the OSRCatchBlock. This was fixed
under FSD by always splitting the block, however, it should also
be applied to other OSR modes as long as OSR exception edges
are expected to be found.

Signed-off-by: Nicholas Coughlin <cnic@ca.ibm.com>