Set array copy flags in VP for recognized methods
VP may not have the type information of the source or destination arrays
we are copying. However for the specialized String compression array
copies we can be certain that both the source and destination arrays are
primitive arrays. As such it is always safe to set the primitive array
copy flags in VP for such recognized methods.

Failure to set these flags can get us in an invalid state in which we
fail to properly generate array store checks or misinterpret the special
String compression array copies as reference array copies, both of which
are incorrect.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>