Create OMR::IlValue and TR::IlValue classes
The IlBuilder class currently typedefs TR::IlValue as
TR::SymbolReference. This commit introduces a fully
fledged TR::IlValue class and defines an OMR::IlValue
base class. It does not use the same extensible class
mechanism as used in the OMR compiler, but it is the
same (somewhat simpler) extension mechanism used in
other IlBuilder (MethodBuilder, TypeDictionary, etc.)
classes. That means most of the functionality being
introduced in this commit will be in OMR::IlValue.

The reason to make TR::IlValue a class in its own
right is to facilitate better quality IL generation
for IlBuilder.

The idea behind this class is to track IlValues initially
as TR::Node pointers. When services consume an IlValue
inside the same basic block, the TR::Node pointer can
be referenced directly. To handle the "same basic block"
question, each OMR::IlValue will store the TR::Block
pointer where the TR::Node pointer was created, and
"consuming" the IlValue will also need to provide the
current TR::Block pointer for comparison.

If the current block is different than the block where
the TR::Node was computed, then that pointer cannot be
directly used. Instead, the node value must be stored
into a local variable. OMR::IlValue will not directly
take care of this storage, but to facilitate the
insertion of the store, OMR::IlValue will store the
current TR::TreeTop pointer being used when the TR::Node
is created. The idea will be that each IlValue will be
anchored by its own TreeTop, at least initially, so
that the ordering of operations is preserved.

By recording the TR::TreeTop for the initial evaluation
of the IlValue, IlValue can give IlBuilder the ability
to rewrite that TreeTop as a store without affecting
the anchor point for the first evaluation of the
TR::Node.

With this kind of scheme, IlValues that are only used
within the same basic block should be able to directly
reference the  node (i.e. correctly commoning the node)
with much less overhead (only one extra TreeTop), whereas
non-local references (i.e. ones in another basic block)
will properly use a local variable to which the node's
value has been stored. This mechanism should substantially
improve the efficiency of IL generation from IlBuilder,
requiring substantially less memory and primarily just
dead trees eliination just to get to the point of having
IL that "looks" like it came from the code the client
wrote/executed.

This commit introduces the new OMR::IlValue and TR::IlValue
classes and modifies primarily the IlBuilder class to
use them. Since JitBuilder was using the CP index field
from symbol references as a "name" for values in the
log output, I also introduced a numbering (ID) scheme
for values that is controlled by MethodBuilder (so the
method builder starts numbering values at zero and
increments for each new value created). The log file
output can be further improved but that's not part
of this commit. A subsequent commit will also adjust
the optimization strategy to remove unneeded opts at
the beginning.

Issue: #834
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>